Moved post metadata-related code into its own file
