+++ app/post.go
@@ -983,103 +983,6 @@ func (a *App) DoPostAction(postId string, actionId string, userId string) *model
 	return nil
 }
 
-func (a *App) PreparePostListForClient(originalList *model.PostList) (*model.PostList, *model.AppError) {
-	list := &model.PostList{
-		Posts: make(map[string]*model.Post),
-		Order: originalList.Order,
-	}
-
-	for id, originalPost := range originalList.Posts {
-		post, err := a.PreparePostForClient(originalPost)
-		if err != nil {
-			return originalList, err
-		}
-
-		list.Posts[id] = post
-	}
-
-	return list, nil
-}
-
-func (a *App) PreparePostForClient(originalPost *model.Post) (*model.Post, *model.AppError) {
-	post := originalPost.Clone()
-
-	// TODO maybe move warnings into here since it might be better to continue on as best we can
-	// if something does go wrong
-	var err *model.AppError
-
-	needReactionCounts := post.ReactionCounts == nil
-	needEmojis := post.Emojis == nil
-	needImageDimensions := post.ImageDimensions == nil
-	needOpenGraphData := post.OpenGraphData == nil
-
-	var reactions []*model.Reaction
-	if needReactionCounts || needEmojis {
-		reactions, err = a.GetReactionsForPost(post.Id)
-		if err != nil {
-			return post, err
-		}
-	}
-
-	if needReactionCounts {
-		post.ReactionCounts = model.CountReactions(reactions)
-	}
-
-	if post.FileInfos == nil {
-		fileInfos, err := a.GetFileInfosForPost(post.Id, false)
-		if err != nil {
-			return post, err
-		}
-
-		post.FileInfos = fileInfos
-	}
-
-	if needEmojis {
-		emojis, err := a.getCustomEmojisForPost(post.Message, reactions)
-		if err != nil {
-			return post, err
-		}
-
-		post.Emojis = emojis
-	}
-
-	post = a.PostWithProxyAddedToImageURLs(post)
-
-	if needImageDimensions || needOpenGraphData {
-		if needImageDimensions {
-			post.ImageDimensions = []*model.PostImageDimensions{}
-		}
-
-		if needOpenGraphData {
-			post.OpenGraphData = []*opengraph.OpenGraph{}
-		}
-
-		// TODO
-	}
-
-	return post, nil
-}
-
-func (a *App) getCustomEmojisForPost(message string, reactions []*model.Reaction) ([]*model.Emoji, *model.AppError) {
-	names := model.EMOJI_PATTERN.FindAllString(message, -1)
-
-	for _, reaction := range reactions {
-		names = append(names, reaction.EmojiName)
-	}
-
-	if len(names) == 0 {
-		return []*model.Emoji{}, nil
-	}
-
-	names = model.RemoveDuplicateStrings(names)
-
-	for i, name := range names {
-		names[i] = strings.Trim(name, ":")
-	}
-
-	return a.GetMultipleEmojiByName(names)
-}
-
 func (a *App) PostWithProxyAddedToImageURLs(post *model.Post) *model.Post {
 	if f := a.ImageProxyAdder(); f != nil {
 		return post.WithRewrittenImageURLs(f)
